Fixing small bugs here and there

+++ app/models/offer.rb
@@ -5,13 +5,13 @@ class Offer < ActiveRecord::Base
   has_and_belongs_to_many :users
 
   def make_swap
-      user1 = self.users[0]
+    user1 = self.users[0]
     user2 = self.users[1]
     user1_products = []
     user2_products = []
     self.products.each do |p|
       p.offers.each do |o|
-        if o != self
+        if o != self && o.completed == 'n'
           #UserMailer.offer_cancelled(@user2, self).deliver
           o.destroy
         end
@@ -26,11 +26,13 @@ class Offer < ActiveRecord::Base
     end
     user1_products.each do |p|
       p.user = user2
+      p.available = 'n'
       p.save
       user2.products << p
     end
     user2_products.each do |p|
       p.user = user1
+      p.available = 'n'
       p.save
       user1.products << p
     end
